fix: Mapped[Decimal] → Mapped[float] 되돌리기 + 주문 정렬 수정
모델 전체에서 Mapped[Decimal]을 Mapped[float]로 되돌림.
3/17 감사에서 Decimal로 변경했으나, 코드 전체가 float 기반이라
add_pending_earnings()에서 Decimal+float TypeError 반복 발생.
DB 컬럼(Numeric)은 그대로 유지하여 정밀도 보장.

주문 히스토리 정렬을 update_time_ms → order_id로 변경하여
시간순 표시 오류 수정.

+++ app/models/backtest_run.py
@@ -1,6 +1,5 @@
 import uuid
 from datetime import datetime
-from decimal import Decimal
 
 from sqlalchemy import BigInteger, CheckConstraint, DateTime, ForeignKey, Index, Numeric, String
 from sqlalchemy.dialects.postgresql import JSONB
@@ -24,7 +23,7 @@ class BacktestRun(CreatedAtMixin, Base):
     combos: Mapped[list[dict] | None] = mapped_column(JSONB, nullable=True)
     strategies: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
     strategy_params: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
-    initial_usdt: Mapped[Decimal] = mapped_column(Numeric, nullable=False)
+    initial_usdt: Mapped[float] = mapped_column(Numeric, nullable=False)
     start_ts_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     end_ts_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     status: Mapped[str] = mapped_column(String, nullable=False, server_default="PENDING")
